improving the MultiForm
if a fixedValue prop is passed then the relevent pair is disabled and the default value same as of fixedValue.

This can be improved further by controlling the defaultValue and isFixed state of the inputBox through different props. this way i can set the default value of the inputBoxed without necessarily disabling it.